requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.7.1
astroid==2.0.4
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==2.1.3
bs4==0.0.1
cached-property==1.5.1
cchardet==2.1.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
colorama==0.3.9
compoundfiles==0.3
configparser==3.5.0
crypto==1.4.1
cryptography==2.2.2
cycler==0.10.0
dask==0.18.1
decorator==4.3.0
defusedxml==0.6.0
dict2xml==1.5
dicttoxml==1.7.4
Django==2.0.7
django-cors-headers==2.4.0
django-extensions==2.1.9
django-pyodbc-azure==2.0.6.1
django-request==1.5.4
django-request-logging==0.6.5
django-sslserver==0.20
djangorestframework==3.7.7
djangorestframework-xml==1.3.0
dnspython==1.16.0
eml-parser==1.10
entrypoints==0.2.3
exchangelib==2.0.0
file-magic==0.4.0
future==0.17.1
gast==0.2.2
get==2019.4.13
grpcio==1.20.1
h5py==2.9.0
html5lib==1.0.1
httplib2==0.11.3
idna==2.6
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isodate==0.6.0
jedi==0.13.3
Jinja2==2.10
json2xml==2.2.0
jsonfield==2.0.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==19.0.2
lazy-object-proxy==1.3.1
livereload==2.5.2
lxml==4.4.1
Markdown==2.6.11
MarkupSafe==1.0
mistune==0.8.4
mkdocs==0.17.5
mock==3.0.5
mysql-connector-python==8.0.16
Naked==0.1.31
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
ntlm-auth==1.4.0
numpy==1.14.5
O365==2.0.1
oauthlib==3.1.0
opencv-python==4.1.0.25
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
Pillow==6.0.0
post==2019.4.13
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.6.0
public==2019.4.13
py4j==0.10.7
pycparser==2.18
pycryptodome==3.6.4
Pygments==2.3.1
PyMySQL==0.9.3
pyodbc==4.0.26
pyrsistent==0.15.1
pyspark==2.4.2
python-dateutil==2.7.3
python-magic==0.4.15
pytz==2018.5
pywin32==224
pywin32-ctypes==0.2.0
pywinpty==0.5.5
pyxlsb==1.0.3
PyYAML==3.13
pyzmq==18.0.1
qtconsole==4.4.4
query-string==2019.4.13
request==2019.4.13
requests==2.18.4
requests-ntlm==1.1.0
requests-oauthlib==1.2.0
scipy==1.3.0
selenium==3.14.0
Send2Trash==1.5.0
shellescape==3.4.1
six==1.11.0
sqlparse==0.3.0
stringcase==1.2.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==4.5.3
traitlets==4.3.2
typed-ast==1.1.0
typing==3.6.4
tzlocal==2.0.0
urllib3==1.22
virtualenv==16.0.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wrapt==1.10.11
xlrd==1.1.0
xmltodict==0.11.0
yagmail==0.11.214
